Abstract data to data file
